-fixed build errors -cleaned routes:created separate files for utils/types
