+++ sim/stores/settings/general/store.ts
@@ -9,6 +9,7 @@ export const useGeneralStore = create<GeneralStore>()(
         isAutoConnectEnabled: true,
         isDebugModeEnabled: false,
         isAutoFillEnvVarsEnabled: true,
+        theme: 'system' as const,
         toggleAutoConnect: () =>
           set((state) => ({
             isAutoConnectEnabled: !state.isAutoConnectEnabled,
@@ -16,6 +17,7 @@ export const useGeneralStore = create<GeneralStore>()(
         toggleDebugMode: () => set((state) => ({ isDebugModeEnabled: !state.isDebugModeEnabled })),
         toggleAutoFillEnvVars: () =>
           set((state) => ({ isAutoFillEnvVarsEnabled: !state.isAutoFillEnvVarsEnabled })),
+        setTheme: (theme) => set(() => ({ theme })),
       }),
       {
         name: 'general-settings',
